Refactor based on E-RAG, change folder name to text2sql
