+++ comps/texttosql/scr/README.md
@@ -1,18 +1,38 @@
-# 🛢🔗 Text-to-SQL Microservice with Langchain
+# 🛢 Text-to-SQL Microservice
 
-This README provides set-up instructions and comprehensive details regarding the Text-to-SQL microservices via LangChain. In this configuration, we will employ PostgresDB as our example database to showcase this microservice.
+In today's data-driven world, the ability to efficiently extract insights from databases is crucial. However, querying databases often requires specialized knowledge of SQL(Structured Query Language) and database schemas, which can be a barrier for non-technical users. This is where the Text-to-SQL microservice comes into play, leveraging the power of LLMs and agentic frameworks to bridge the gap between human language and database queries. This microservice is built on LangChain/LangGraph frameworks.
+
+The microservice enables a wide range of use cases, making it a versatile tool for businesses, researchers, and individuals alike. Users can generate queries based on natural language questions, enabling them to quickly retrieve relevant data from their databases. Additionally, the service can be integrated into ChatBots, allowing for natural language interactions and providing accurate responses based on the underlying data. Furthermore, it can be utilized to build custom dashboards, enabling users to visualize and analyze insights based on their specific requirements, all through the power of natural language.
 
 ---
 
-## 🚀 Start Microservice with Python（Option 1）
+## 🛠️ Features
 
-### Install Requirements
+**Implement SQL Query based on input text**: Transform user-provided natural language into SQL queries, subsequently executing them to retrieve data from SQL databases.
+
+---
+
+## ⚙️ Implementation
+
+The text-to-sql microservice able to implement with various framework and support various types of SQL databases.
+
+### 🔗 Utilizing Text-to-SQL with Langchain framework
+
+
+
+The follow guide provides set-up instructions and comprehensive details regarding the Text-to-SQL microservices via LangChain. In this configuration, we will employ PostgresDB as our example database to showcase this microservice.
+
+---
+
+#### 🚀 Start Microservice with Python（Option 1）
+
+#### Install Requirements
 
 ```bash
 pip install -r requirements.txt
 ```
 
-### Start PostgresDB Service
+#### Start PostgresDB Service
 
 We will use [Chinook](https://github.com/lerocha/chinook-database) sample database as a default to test the Text-to-SQL microservice. Chinook database is a sample database ideal for demos and testing ORM tools targeting single and multiple database servers.
 
@@ -21,12 +41,12 @@ export POSTGRES_USER=postgres
 export POSTGRES_PASSWORD=testpwd
 export POSTGRES_DB=chinook
 
-cd comps/texttosql/langchain
+cd comps/texttosql
 
 docker run --name postgres-db --ipc=host -e POSTGRES_USER=${POSTGRES_USER} -e POSTGRES_HOST_AUTH_METHOD=trust -e POSTGRES_DB=${POSTGRES_DB} -e POSTGRES_PASSWORD=${POSTGRES_PASSWORD} -p 5442:5432 -d -v ./chinook.sql:/docker-entrypoint-initdb.d/chinook.sql postgres:latest
 ```
 
-### Start TGI Service
+#### Start TGI Service
 
 ```bash
 export HUGGINGFACEHUB_API_TOKEN=${HUGGINGFACEHUB_API_TOKEN}
@@ -36,7 +56,7 @@ export TGI_PORT=8008
 docker run -d --name="texttosql-tgi-endpoint" --ipc=host -p $TGI_PORT:80 -v ./data:/data --shm-size 1g -e HF_TOKEN=${HUGGINGFACEHUB_API_TOKEN} -e model=${LLM_MODEL_ID} ghcr.io/huggingface/text-generation-inference:2.1.0 --model-id $LLM_MODEL_ID
 ```
 
-### Verify the TGI Service
+#### Verify the TGI Service
 
 ```bash
 export your_ip=$(hostname -I | awk '{print $1}')
@@ -46,39 +66,39 @@ curl http://${your_ip}:${TGI_PORT}/generate \
   -H 'Content-Type: application/json'
 ```
 
-### Setup Environment Variables
+#### Setup Environment Variables
 
 ```bash
 export TGI_LLM_ENDPOINT="http://${your_ip}:${TGI_PORT}"
 ```
 
-### Start Text-to-SQL Microservice with Python Script
+#### Start Text-to-SQL Microservice with Python Script
 
 Start Text-to-SQL microservice with below command.
 
 ```bash
-python3 main.py
+python3 texttosql_microservice.py
 ```
 
 ---
 
-## 🚀 Start Microservice with Docker (Option 2)
+### 🚀 Start Microservice with Docker (Option 2)
 
-### Start PostGreSQL Database Service
+#### Start PostGreSQL Database Service
 
 Please refer to section [Start PostgresDB Service](#start-postgresdb-service)
 
-### Start TGI Service
+#### Start TGI Service
 
 Please refer to section [Start TGI Service](#start-tgi-service)
 
-### Setup Environment Variables
+#### Setup Environment Variables
 
 ```bash
 export TGI_LLM_ENDPOINT="http://${your_ip}:${TGI_PORT}"
 ```
 
-### Build Docker Image
+#### Build Docker Image
 
 ```bash
 cd GenAIComps/
@@ -114,7 +134,7 @@ docker run  --runtime=runc --name="comps-langchain-texttosql"  -p 9090:8080 --ip
 
 ---
 
-## ✅ Invoke the microservice.
+### ✅ Invoke the microservice.
 
 The Text-to-SQL microservice exposes the following API endpoints:
