+++ comps/texttosql/scr/requirements.txt
